clean up board container and board
