+++ app/components/GameView.jsx
@@ -16,23 +16,22 @@ export default class GameView extends Component {
     allWords.on('value', snapshot => {
       const wordArray = snapshot.val()
       this.setState({allWords: wordArray})
-      console.log('~~~~~~~~~~~~~', this.state)
     })
     const currentGameStatus = firebase.database().ref(`gameInstances/${this.props.params.gameId}/currentGameStatus`)
     currentGameStatus.on('value', snapshot => {
       this.setState({currentGameStatus: snapshot.val()})
-      console.log('!!!!!!!!!!!!!!!!!!', Object.keys(this.state).length, this.state)
     })
   }
 
   render() {
     return (
       <div>
-        {Object.keys(this.state).length >= 2 &&
+        {
+          Object.keys(this.state).length >= 2 &&
           <BoardContainer
-          allWords={this.state.allWords}
-          currentGameStatus={this.state.currentGameStatus}
-          gameId={this.props.params.gameId}
+            allWords={this.state.allWords}
+            currentGameStatus={this.state.currentGameStatus}
+            gameId={this.props.params.gameId}
           />
         }
       </div>
